Last touches and rename
